[Issue-46] post 관련 모델 수정 및 추가

+++ Picon/app/src/main/java/com/yapp/picon/data/model/Coordinate.kt
@@ -1,6 +1,8 @@
 package com.yapp.picon.data.model
 
+import java.math.BigDecimal
+
 data class Coordinate(
-    val lat: Double,
-    val lng: Double
+    val lat: BigDecimal,
+    val lng: BigDecimal
 )
